backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
defusedxml==0.7.1
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
fastapi==0.115.12
filetype==1.2.0
flake8==7.3.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
line_profiler==4.2.0
Markdown==3.8
mccabe==0.7.0
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.4.0
PyJWT==2.9.0
python-dotenv==1.1.0
python3-openid==3.2.0
reportlab==4.4.2
requests==2.32.4
requests-oauthlib==2.0.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
